Add build.bat and a hack for llvm_lto_status_t
llvm_lto_status_t is a little tricky. It is an enum that has forward
declaration and in-place declaration. To discover and associate forward
declarations, we visit the TU again, arrive at the cursor and the
immediately next typedef decl is the spelling we're looking for. In this
case, it is not true, because the in-place decl wins. This is a
complicated case as the code is written right now, and since shipping is
a feature, need to place this terrible hack in.