First step to OOPifying LLVMSharp
Now the structs have instance methods. This is a first step to organize
the code, but really the approach is to get full objects. We'll have to
keep this approach around though, because some folks will want to not
pay object allocation costs, so for them this is the OOP API, but most
people will want the "FULL" experience with types for each instruction.